extracting copyright from signin signup and newproperty into new component
