+++ client/src/components/Property/PropertyForm/PropertyGalleryForm.js
@@ -3,7 +3,7 @@ import FileUpload from "./FileUpload";
 const PropertyGalleryForm = () => {
   return (
     <>
-      {/* <FileUpload /> */}
+      <FileUpload />
     </>
   );
 };
